fix: add cli test
